lots of CoinDetail Model changes

+++ app/src/main/java/com/example/coinapiapp/data/data_source/dtoModel/CoinDetailDTO/ConvertedVolume.kt
@@ -3,5 +3,5 @@ package com.example.coinapiapp.data.data_source.dtoModel.CoinDetailDTO
 data class ConvertedVolume(
     val btc: Double,
     val eth: Double,
-    val usd: Int
+    val usd: Double
 )
